chore: rename studio plugin
