+++ packages/astro-plugin-studio/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@pandacss/vite-plugin-studio",
+  "name": "@pandacss/astro-plugin-studio",
   "version": "0.20.1",
   "description": "Vite plugin for Pandacss Studio",
   "author": "Segun Adebayo <joseshegs@gmail.com>",
