+++ packages/astro-plugin-studio/tsconfig.json
