+++ packages/studio/astro.config.mjs
@@ -1,5 +1,5 @@
 import react from '@astrojs/react'
-import panda from '@pandacss/vite-plugin-studio'
+import panda from '@pandacss/astro-plugin-studio'
 import { defineConfig } from 'astro/config'
 
 // https://astro.build/config
